Create Administration namespace
