+++ lib/bayberry_web/controllers/administration_controller.ex
@@ -1,0 +1,7 @@
+defmodule BayberryWeb.AdministrationController do
+  use BayberryWeb, :controller
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+end
