+++ lib/bayberry_web/templates/administration/index.html.eex
@@ -4,5 +4,5 @@
   <li><%= link "Users", to: accounts_user_path(@conn, :index) %></li>
   <li><%= link "Authors", to: blog_author_path(@conn, :index) %></li>
   <li><%= link "Articles", to: blog_article_path(@conn, :index) %></li>
-  <li><%= link "Visitors", to: accounts_administration_path(@conn, :visitors) %></li>
+  <li><%= link "Visitors", to: administration_visitor_path(@conn, :index) %></li>
 </ul>
