+++ lib/bayberry_web/views/administration_view.ex
@@ -1,0 +1,3 @@
+defmodule BayberryWeb.AdministrationView do
+  use BayberryWeb, :view
+end
